+++ cc.config.json
@@ -216,10 +216,10 @@
                 "cocos/spine/index.ts": "cocos/spine/index.jsb.ts",
                 "cocos/dragon-bones/index.ts": "cocos/dragon-bones/index.jsb.ts",
                 "cocos/physics/physx/instantiate.ts": "cocos/physics/physx/instantiate.jsb.ts",
-                "cocos/core/scene-graph/node.ts": "cocos/core/scene-graph/node.jsb.ts",
-                "cocos/core/scene-graph/scene.ts": "cocos/core/scene-graph/scene.jsb.ts",
-                "cocos/core/scene-graph/scene-globals.ts": "cocos/core/scene-graph/scene-globals.jsb.ts",
-                "cocos/core/scene-graph/index.ts": "cocos/core/scene-graph/index.jsb.ts",
+                "cocos/scene-graph/node.ts": "cocos/scene-graph/node.jsb.ts",
+                "cocos/scene-graph/scene.ts": "cocos/scene-graph/scene.jsb.ts",
+                "cocos/scene-graph/scene-globals.ts": "cocos/scene-graph/scene-globals.jsb.ts",
+                "cocos/scene-graph/index.ts": "cocos/scene-graph/index.jsb.ts",
                 "cocos/core/root.ts": "cocos/core/root.jsb.ts",
                 "cocos/asset/asset-manager/builtin-res-mgr.ts": "cocos/asset/asset-manager/builtin-res-mgr.jsb.ts",
                 "cocos/asset/assets/index.ts": "cocos/asset/assets/index.jsb.ts",
