+++ cocos/misc/index.ts
@@ -1,5 +1,5 @@
 /*
- Copyright (c) 2017-2020 Xiamen Yaji Software Co., Ltd.
+ Copyright (c) 2018-2020 Xiamen Yaji Software Co., Ltd.
 
  http://www.cocos.com
 
@@ -23,17 +23,10 @@
  THE SOFTWARE.
 */
 
-import './node-event-processor';
-import './deprecated';
-import './deprecated-3.7.0';
-// @deprecated since v3.7, please use Node instead.
-export { Node } from './node';
-export { Node as BaseNode } from './node'; //reserve BaseNode for compatibility. Note: should export it after export Node.
-export { Scene } from './scene';
-export { Layers } from './layers';
-export { find } from './find';
-export * from './node-enum';
-export * from './node-event';
-export * from './scene-globals';
+import './component-event-handler.schema';
+
+export { Camera } from './camera-component';
+export { ModelRenderer } from './model-renderer';
+export { Renderer } from './renderer';
+/** deprecated */
 export * from './deprecated';
-export { default as NodeActivator } from './node-activator';
