+++ cocos/misc/renderer.ts
@@ -27,10 +27,10 @@ import { EDITOR } from 'internal:constants';
 import {
     ccclass, type, displayOrder, displayName, disallowMultiple,
 } from 'cc.decorator';
-import { Material } from '../../asset/assets/material';
-import { Component } from './component';
-import { IMaterialInstanceInfo, MaterialInstance } from '../../render-scene/core/material-instance';
-import { warnID } from '../platform/debug';
+import { Material } from '../asset/assets/material';
+import { Component } from '../scene-graph';
+import { IMaterialInstanceInfo, MaterialInstance } from '../render-scene/core/material-instance';
+import { warnID } from '../core/platform/debug';
 
 const _matInsInfo: IMaterialInstanceInfo = {
     parent: null!,
