+++ cocos/scene-graph/category.json
