+++ cocos/scene-graph/node-dev.ts
@@ -24,11 +24,11 @@
 */
 
 import { EDITOR, DEV, TEST } from 'internal:constants';
-import { CCObject } from '../data/object';
-import * as js from '../utils/js';
-import { legacyCC } from '../global-exports';
-import { error, errorID, getError } from '../platform/debug';
-import { Component } from '../components';
+import { CCObject } from '../core/data/object';
+import * as js from '../core/utils/js';
+import { legacyCC } from '../core/global-exports';
+import { error, errorID, getError } from '../core/platform/debug';
+import { Component } from './component';
 
 const Destroying = CCObject.Flags.Destroying;
