+++ cocos/scene-graph/node-event-processor.ts
@@ -23,14 +23,14 @@
  THE SOFTWARE.
 */
 
-import { CallbacksInvoker } from '../event/callbacks-invoker';
-import { Event, EventMouse, EventTouch, Touch } from '../../input/types';
-import { Vec2 } from '../math/vec2';
+import { CallbacksInvoker } from '../core/event/callbacks-invoker';
+import { Event, EventMouse, EventTouch, Touch } from '../input/types';
+import { Vec2 } from '../core/math/vec2';
 import { Node } from './node';
-import { legacyCC } from '../global-exports';
-import { Component } from '../components/component';
+import { legacyCC } from '../core/global-exports';
+import { Component } from './component';
 import { NodeEventType } from './node-event';
-import { InputEventType, SystemEventTypeUnion } from '../../input/types/event-enum';
+import { InputEventType, SystemEventTypeUnion } from '../input/types/event-enum';
 
 const _cachedArray = new Array<Node>(16);
 let _currentHovered: Node | null = null;
