+++ cocos/scene-graph/node.ts
@@ -27,22 +27,22 @@ import { ccclass, editable, serializable, type } from 'cc.decorator';
 import { DEV, DEBUG, EDITOR } from 'internal:constants';
 import { Layers } from './layers';
 import { NodeUIProperties } from './node-ui-properties';
-import { legacyCC } from '../global-exports';
+import { legacyCC } from '../core/global-exports';
 import { nodePolyfill } from './node-dev';
-import { ISchedulable } from '../scheduler';
-import { approx, EPSILON, Mat3, Mat4, Quat, Vec3 } from '../math';
+import { ISchedulable } from '../core/scheduler';
+import { approx, EPSILON, Mat3, Mat4, Quat, Vec3 } from '../core/math';
 import { NodeSpace, TransformBit } from './node-enum';
-import { CustomSerializable, editorExtrasTag, SerializationContext, SerializationOutput, serializeTag } from '../data';
-import { errorID, warnID, error, log, getError } from '../platform/debug';
-import { Component } from '../components/component';
-import { property } from '../data/decorators/property';
-import { CCObject } from '../data/object';
-import * as js from '../utils/js';
-import IdGenerator from '../utils/id-generator';
+import { CustomSerializable, editorExtrasTag, SerializationContext, SerializationOutput, serializeTag } from '../core/data';
+import { errorID, warnID, error, log, getError } from '../core/platform/debug';
+import { Component } from './component';
+import { property } from '../core/data/decorators/property';
+import { CCObject } from '../core/data/object';
+import * as js from '../core/utils/js';
+import IdGenerator from '../core/utils/id-generator';
 import type { Scene } from './scene';
-import { PrefabInfo } from '../utils/prefab/prefab-info';
+import { PrefabInfo } from '../core/utils/prefab/prefab-info';
 import { NodeEventType } from './node-event';
-import { Event } from '../../input/types';
+import { Event } from '../input/types';
 
 const Destroying = CCObject.Flags.Destroying;
 const DontDestroy = CCObject.Flags.DontDestroy;
