+++ cocos/scene-graph/scene-globals.ts
@@ -23,19 +23,19 @@
 import { ccclass, visible, type, displayOrder, readOnly, slide, range, rangeStep,
     editable, serializable, rangeMin, tooltip, formerlySerializedAs, displayName } from 'cc.decorator';
 import { BAIDU } from 'internal:constants';
-import { TextureCube } from '../../asset/assets/texture-cube';
-import { CCFloat, CCInteger } from '../data/utils/attribute';
-import { Color, Quat, Vec3, Vec2, Vec4 } from '../math';
-import { Ambient } from '../../render-scene/scene/ambient';
-import { Shadows, ShadowType, ShadowSize } from '../../render-scene/scene/shadows';
-import { Skybox, EnvironmentLightingType } from '../../render-scene/scene/skybox';
-import { Octree } from '../../render-scene/scene/octree';
-import { Fog, FogType } from '../../render-scene/scene/fog';
+import { TextureCube } from '../asset/assets/texture-cube';
+import { CCFloat, CCInteger } from '../core/data/utils/attribute';
+import { Color, Quat, Vec3, Vec2, Vec4 } from '../core/math';
+import { Ambient } from '../render-scene/scene/ambient';
+import { Shadows, ShadowType, ShadowSize } from '../render-scene/scene/shadows';
+import { Skybox, EnvironmentLightingType } from '../render-scene/scene/skybox';
+import { Octree } from '../render-scene/scene/octree';
+import { Fog, FogType } from '../render-scene/scene/fog';
 import { Node } from './node';
-import { legacyCC } from '../global-exports';
-import { Root } from '../root';
-import { warnID } from '../platform/debug';
-import { Material } from '../../asset/assets/material';
+import { legacyCC } from '../core/global-exports';
+import { Root } from '../core/root';
+import { warnID } from '../core/platform/debug';
+import { Material } from '../asset/assets/material';
 
 const _up = new Vec3(0, 1, 0);
 const _v3 = new Vec3();
