+++ exports/base.ts
@@ -33,12 +33,13 @@ import '../cocos/core/legacy';
 import * as renderer from '../cocos/render-scene';
 import * as gfx from '../cocos/gfx';
 
-
 // LOAD ENGINE CORE
 export * from '../cocos/core';
 
 export * from '../cocos/rendering';
 export * from '../cocos/rendering/custom/builtin-pipelines';
+export * from '../cocos/scene-graph';
+export * from '../cocos/misc';
 
 export { gfx };
